+++ src/core.rs
@@ -1,18 +1,16 @@
-pub(crate) use self::docker::Docker;
+pub(crate) use container::Docker;
+pub(crate) use container_async::DockerAsync;
+
 pub use self::{
     container::Container,
     container_async::ContainerAsync,
-    docker::RunArgs,
-    image::{Image, Port, WaitFor},
+    image::{Image, Port, RunnableImage, WaitFor},
 };
 
 mod container;
 mod container_async;
-mod docker;
 pub mod env;
 mod image;
 
 pub(crate) mod logs;
 pub(crate) mod ports;
-
-pub(crate) use container_async::DockerAsync;