+++ src/images/dynamodb_local.rs
@@ -1,7 +1,6 @@
 use crate::{core::WaitFor, Image};
-use std::collections::HashMap;
 
-const CONTAINER_IDENTIFIER: &str = "amazon/dynamodb-local";
+const NAME: &str = "amazon/dynamodb-local";
 const DEFAULT_WAIT: u64 = 2000;
 const DEFAULT_TAG: &str = "latest";
 
@@ -34,12 +33,13 @@ impl Default for DynamoDb {
 
 impl Image for DynamoDb {
     type Args = DynamoDbArgs;
-    type EnvVars = HashMap<String, String>;
-    type Volumes = HashMap<String, String>;
-    type EntryPoint = std::convert::Infallible;
 
-    fn descriptor(&self) -> String {
-        format!("{}:{}", CONTAINER_IDENTIFIER, &self.tag)
+    fn name(&self) -> String {
+        NAME.to_owned()
+    }
+
+    fn tag(&self) -> String {
+        self.tag.clone()
     }
 
     fn ready_conditions(&self) -> Vec<WaitFor> {
@@ -49,30 +49,5 @@ impl Image for DynamoDb {
             ),
             WaitFor::millis(DEFAULT_WAIT),
         ]
-    }
-
-    fn args(&self) -> <Self as Image>::Args {
-        self.arguments.clone()
-    }
-
-    fn volumes(&self) -> Self::Volumes {
-        HashMap::new()
-    }
-
-    fn env_vars(&self) -> Self::EnvVars {
-        HashMap::new()
-    }
-
-    fn with_args(self, arguments: <Self as Image>::Args) -> Self {
-        DynamoDb { arguments, ..self }
-    }
-}
-
-impl DynamoDb {
-    pub fn with_tag(self, tag_str: &str) -> Self {
-        DynamoDb {
-            tag: tag_str.to_string(),
-            ..self
-        }
     }
 }