+++ src/images/hello_world.rs
@@ -1,36 +1,20 @@
 use crate::{core::WaitFor, Image};
-use std::collections::HashMap;
 
 #[derive(Debug, Default)]
 pub struct HelloWorld;
 
 impl Image for HelloWorld {
     type Args = Vec<String>;
-    type EnvVars = HashMap<String, String>;
-    type Volumes = HashMap<String, String>;
-    type EntryPoint = std::convert::Infallible;
 
-    fn descriptor(&self) -> String {
-        String::from("hello-world")
+    fn name(&self) -> String {
+        "hello-world".to_owned()
+    }
+
+    fn tag(&self) -> String {
+        "latest".to_owned()
     }
 
     fn ready_conditions(&self) -> Vec<WaitFor> {
         vec![WaitFor::message_on_stdout("Hello from Docker!")]
-    }
-
-    fn args(&self) -> <Self as Image>::Args {
-        vec![]
-    }
-
-    fn volumes(&self) -> Self::Volumes {
-        HashMap::new()
-    }
-
-    fn env_vars(&self) -> Self::EnvVars {
-        HashMap::new()
-    }
-
-    fn with_args(self, _: <Self as Image>::Args) -> Self {
-        self
     }
 }