+++ src/images/kafka.rs
@@ -1,7 +1,7 @@
 use crate::{core::WaitFor, Image};
 use std::collections::HashMap;
 
-const CONTAINER_IDENTIFIER: &str = "confluentinc/cp-kafka";
+const NAME: &str = "confluentinc/cp-kafka";
 const DEFAULT_TAG: &str = "6.1.1";
 
 pub const KAFKA_PORT: u16 = 9093;
@@ -39,15 +39,6 @@ pub struct Kafka {
     arguments: KafkaArgs,
     env_vars: HashMap<String, String>,
     tag: String,
-}
-
-impl Kafka {
-    pub fn with_tag<T: Into<String>>(self, tag: T) -> Self {
-        Self {
-            tag: tag.into(),
-            ..self
-        }
-    }
 }
 
 impl Default for Kafka {
@@ -93,31 +84,20 @@ impl Default for Kafka {
 
 impl Image for Kafka {
     type Args = KafkaArgs;
-    type EnvVars = HashMap<String, String>;
-    type Volumes = HashMap<String, String>;
-    type EntryPoint = std::convert::Infallible;
 
-    fn descriptor(&self) -> String {
-        format!("{}:{}", CONTAINER_IDENTIFIER, &self.tag)
+    fn name(&self) -> String {
+        NAME.to_owned()
+    }
+
+    fn tag(&self) -> String {
+        self.tag.clone()
     }
 
     fn ready_conditions(&self) -> Vec<WaitFor> {
         vec![WaitFor::message_on_stdout("Creating new log file")]
     }
 
-    fn args(&self) -> <Self as Image>::Args {
-        self.arguments.clone()
-    }
-
-    fn env_vars(&self) -> Self::EnvVars {
-        self.env_vars.clone()
-    }
-
-    fn volumes(&self) -> Self::Volumes {
-        HashMap::new()
-    }
-
-    fn with_args(self, arguments: <Self as Image>::Args) -> Self {
-        Self { arguments, ..self }
+    fn env_vars(&self) -> Box<dyn Iterator<Item = (&String, &String)> + '_> {
+        Box::new(self.env_vars.iter())
     }
 }