+++ src/images/mongo.rs
@@ -1,7 +1,6 @@
 use crate::{core::WaitFor, Image};
-use std::collections::HashMap;
 
-const CONTAINER_IDENTIFIER: &str = "mongo";
+const NAME: &str = "mongo";
 const DEFAULT_TAG: &str = "4.0.17";
 
 #[derive(Debug, Default, Clone)]
@@ -33,42 +32,18 @@ impl Default for Mongo {
 
 impl Image for Mongo {
     type Args = MongoArgs;
-    type EnvVars = HashMap<String, String>;
-    type Volumes = HashMap<String, String>;
-    type EntryPoint = std::convert::Infallible;
 
-    fn descriptor(&self) -> String {
-        format!("{}:{}", CONTAINER_IDENTIFIER, &self.tag)
+    fn name(&self) -> String {
+        NAME.to_owned()
+    }
+
+    fn tag(&self) -> String {
+        self.tag.clone()
     }
 
     fn ready_conditions(&self) -> Vec<WaitFor> {
         vec![WaitFor::message_on_stdout(
             "waiting for connections on port",
         )]
-    }
-
-    fn args(&self) -> <Self as Image>::Args {
-        self.arguments.clone()
-    }
-
-    fn env_vars(&self) -> Self::EnvVars {
-        HashMap::new()
-    }
-
-    fn volumes(&self) -> Self::Volumes {
-        HashMap::new()
-    }
-
-    fn with_args(self, arguments: <Self as Image>::Args) -> Self {
-        Mongo { arguments, ..self }
-    }
-}
-
-impl Mongo {
-    pub fn with_tag(self, tag_str: &str) -> Self {
-        Mongo {
-            tag: tag_str.to_string(),
-            ..self
-        }
     }
 }