+++ src/images/orientdb.rs
@@ -1,7 +1,7 @@
 use crate::{core::WaitFor, Image};
 use std::collections::HashMap;
 
-const CONTAINER_IDENTIFIER: &str = "orientdb";
+const NAME: &str = "orientdb";
 const DEFAULT_TAG: &str = "3.1.3";
 
 #[derive(Debug, Default, Clone)]
@@ -38,45 +38,20 @@ impl Default for OrientDb {
 
 impl Image for OrientDb {
     type Args = OrientDbArgs;
-    type EnvVars = HashMap<String, String>;
-    type Volumes = HashMap<String, String>;
-    type EntryPoint = std::convert::Infallible;
 
-    fn descriptor(&self) -> String {
-        format!("{}:{}", CONTAINER_IDENTIFIER, &self.tag)
+    fn name(&self) -> String {
+        NAME.to_owned()
+    }
+
+    fn tag(&self) -> String {
+        self.tag.clone()
     }
 
     fn ready_conditions(&self) -> Vec<WaitFor> {
         vec![WaitFor::message_on_stderr("OrientDB Studio available at")]
     }
 
-    fn args(&self) -> <Self as Image>::Args {
-        self.arguments.clone()
-    }
-
-    fn env_vars(&self) -> Self::EnvVars {
-        self.env_vars.clone()
-    }
-
-    fn volumes(&self) -> Self::Volumes {
-        HashMap::new()
-    }
-
-    fn with_args(self, arguments: <Self as Image>::Args) -> Self {
-        OrientDb { arguments, ..self }
-    }
-}
-
-impl OrientDb {
-    pub fn with_tag(self, tag_str: &str) -> Self {
-        OrientDb {
-            tag: tag_str.to_string(),
-            ..self
-        }
-    }
-
-    pub fn with_env_var<K: Into<String>, V: Into<String>>(mut self, key: K, value: V) -> Self {
-        self.env_vars.insert(key.into(), value.into());
-        self
+    fn env_vars(&self) -> Box<dyn Iterator<Item = (&String, &String)> + '_> {
+        Box::new(self.env_vars.iter())
     }
 }