+++ src/images/redis.rs
@@ -1,7 +1,6 @@
 use crate::{core::WaitFor, Image};
-use std::collections::HashMap;
 
-const CONTAINER_IDENTIFIER: &str = "redis";
+const NAME: &str = "redis";
 const DEFAULT_TAG: &str = "5.0";
 
 #[derive(Debug, Default, Clone)]
@@ -33,40 +32,16 @@ impl Default for Redis {
 
 impl Image for Redis {
     type Args = RedisArgs;
-    type EnvVars = HashMap<String, String>;
-    type Volumes = HashMap<String, String>;
-    type EntryPoint = std::convert::Infallible;
 
-    fn descriptor(&self) -> String {
-        format!("{}:{}", CONTAINER_IDENTIFIER, &self.tag)
+    fn name(&self) -> String {
+        NAME.to_owned()
+    }
+
+    fn tag(&self) -> String {
+        self.tag.clone()
     }
 
     fn ready_conditions(&self) -> Vec<WaitFor> {
         vec![WaitFor::message_on_stdout("Ready to accept connections")]
-    }
-
-    fn args(&self) -> <Self as Image>::Args {
-        self.arguments.clone()
-    }
-
-    fn volumes(&self) -> Self::Volumes {
-        HashMap::new()
-    }
-
-    fn env_vars(&self) -> Self::EnvVars {
-        HashMap::new()
-    }
-
-    fn with_args(self, arguments: <Self as Image>::Args) -> Self {
-        Redis { arguments, ..self }
-    }
-}
-
-impl Redis {
-    pub fn with_tag(self, tag_str: &str) -> Self {
-        Redis {
-            tag: tag_str.to_string(),
-            ..self
-        }
     }
 }